# core/src/main/java/com/muravlev/core/dto/PostDTO.java
package com.muravlev.core.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import java.time.LocalDateTime;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class PostDTO {
    private Long id;
    private String content;
    private String imageUrl;
    private Long authorId;
    private Long channelId;
    private LocalDateTime creationTime;
    private String authorName;
}